fixup! feat: improve merge behavior
